+++ packages/components/src/systems/PatientMedHistory/PatientMedHistoryTable.tsx
@@ -4,7 +4,7 @@ import { Treatment } from '@photonhealth/sdk/dist/types';
 import { IconButton } from '../../particles/IconButton';
 import clsx from 'clsx';
 import { MedHistoryPrescription } from './index';
-import { usePrescribeOptional } from '../PrescribeProvider';
+import { useDraftPrescriptionsOptional } from '../DraftPrescriptions';
 
 export type MedHistoryRowItem = {
   treatment: Treatment;
@@ -24,7 +24,7 @@ export type PatientMedHistoryTableProps = {
 export default function PatientMedHistoryTable(props: PatientMedHistoryTableProps) {
   // Component can be used as a standalone element
   // so PrescribeProvider is not guaranteed to be rendered
-  const prescribeContext = usePrescribeOptional();
+  const draftPrescriptionsContext = useDraftPrescriptionsOptional();
 
   const [isCreatingPrescriptionId, setIsCreatingPrescriptionId] = createSignal<string | undefined>(
     undefined
@@ -48,12 +48,12 @@ export default function PatientMedHistoryTable(props: PatientMedHistoryTableProp
     if (isCreatingPrescriptionId() === undefined) {
       setIsCreatingPrescriptionId(prescription.id);
 
-      if (!prescribeContext) {
-        throw new Error('Refill requires <PrescribeProvider>');
+      if (!draftPrescriptionsContext) {
+        throw new Error('Refill requires <DraftPrescriptionsProvider>');
       }
 
       try {
-        await prescribeContext.tryCreatePrescription({
+        await draftPrescriptionsContext.tryCreatePrescription({
           ...prescription,
           treatment,
           diagnoseCodes: prescription.diagnoses?.map((diagnosis) => diagnosis.code) || []
