+++ packages/components/src/systems/PrescribeProvider/PrescribeProvider.tsx
@@ -8,22 +8,13 @@ import {
   useContext
 } from 'solid-js';
 import { usePhotonClient } from '../SDKProvider';
+import { Prescription } from '@photonhealth/sdk/dist/types';
 import {
-  MutationCreatePrescriptionsArgs,
-  Prescription,
-  PrescriptionState
-} from '@photonhealth/sdk/dist/types';
-import {
-  CreatePrescription,
-  CreatePrescriptions,
-  CreatePrescriptionTemplate,
   GenerateCoverageOptions,
   GetPatientPreferredPharmaciesAndAddress,
-  GetPharmaciesQuery,
-  GetPrescription,
-  UpdatePrescriptionStates
+  GetPharmaciesQuery
 } from '../../fetch';
-import { triggerToast, useRecentOrders } from '../../index';
+import { triggerToast } from '../../index';
 import { useDraftPrescriptions } from '../DraftPrescriptions';
 import {
   combineAllRoutingConstraints,
@@ -33,7 +24,7 @@ import {
 import { createStore } from 'solid-js/store';
 import getLocation from '../../utils/getLocations';
 import { useGoogleService } from '../GoogleServiceProvider';
-import { usePrescribeEventDispatch } from '../PrescribeEventDispatchProvider';
+
 // The order form data will consist of, at least, the list of selected prescription IDs and pharmacy ID.
 // The prescription form data (todo) will consist of a single prescription's data during user input
 // Note: Multiple prescription "sub" forms can be opened/completed within a single order form
@@ -43,7 +34,6 @@ interface PrescribeOrderFormData {
 
 export type PrescribeContextType = {
   // values
-  isLoadingPrefills: Accessor<boolean>;
   coverageOptions: Accessor<CoverageOption[]>;
   routingConstraints: Accessor<RoutingConstraint[]>;
   combinedRoutingConstraint: Accessor<RoutingConstraint>;
@@ -52,12 +42,6 @@ export type PrescribeContextType = {
   selectedCoverageOption: Accessor<CoverageOption | undefined>;
 
   // actions
-  deletePrescription: (id: string) => void;
-  tryCreatePrescription: (
-    prescriptionFormData: PrescriptionFormData,
-    options?: TryCreatePrescriptionTemplateOptions
-  ) => Promise<Prescription>;
-  tryUpdatePrescriptionStates: (ids: string[], state: PrescriptionState) => Promise<boolean>;
   selectOtherCoverageOption: (value: CoverageOption) => void;
   setOrderFormData: <K extends keyof PrescribeOrderFormData>(
     key: K,
@@ -67,68 +51,46 @@ export type PrescribeContextType = {
 
 const PrescribeContext = createContext<PrescribeContextType>();
 
-export type TemplateOverrides = {
-  [key: string]: {
-    daysSupply?: number;
-    dispenseAsWritten?: boolean;
-    dispenseQuantity?: number;
-    dispenseUnit?: string;
-    fillsAllowed?: number;
-    instructions?: string;
-    notes?: string;
-    externalId?: string;
-  };
-};
-
-export type PrescriptionFormData = {
-  id?: string;
-  doNotFillBeforeDate?: string;
-  treatment: {
-    id: string;
-    name: string;
-  };
-  dispenseAsWritten: boolean;
-  dispenseQuantity?: number;
-  dispenseUnit?: string;
-  daysSupply?: number;
-  instructions: string;
-  notes: string;
-  fillsAllowed?: number;
-  catalogId?: string;
-  externalId?: string;
-  diagnoseCodes: string[];
-};
-
 interface PrescribeProviderProps {
   children: JSXElement;
-  templateIdsPrefill: string[];
-  templateOverrides: TemplateOverrides;
-  prescriptionIdsPrefill: string[];
   patientId: string;
   enableCombineAndDuplicate: boolean;
   enableCoverageCheck: boolean;
 }
 
-const transformPrescriptionFormData = (prescription: PrescriptionFormData, patientId: string) => ({
-  externalId: prescription.externalId,
-  patientId: patientId,
-  treatmentId: prescription.treatment?.id,
-  dispenseAsWritten: prescription.dispenseAsWritten,
-  dispenseQuantity: prescription.dispenseQuantity,
-  dispenseUnit: prescription.dispenseUnit,
-  fillsAllowed: prescription.fillsAllowed,
-  daysSupply: prescription.daysSupply,
-  instructions: prescription.instructions,
-  notes: prescription.notes,
-  doNotFillBeforeDate: prescription.doNotFillBeforeDate,
-  diagnoses: prescription.diagnoseCodes
-});
+export type PatientPreferredPharmacy = {
+  id: string;
+  name: string;
+};
+
+export type CoverageOption = {
+  daysSupply: number;
+  dispenseQuantity: number;
+  dispenseUnit: string;
+  id: string;
+  isAlternative: boolean;
+  paRequired: boolean;
+  prescriptionId: string;
+  price: number | null;
+  status: 'COVERED' | 'COVERED_WITH_RESTRICTIONS' | 'NOT_COVERED';
+  statusMessage: string;
+  treatment: { id: string; name: string };
+  alerts: Array<{ label: string; text: string }>;
+  pharmacy: { id: string; name: string };
+};
+
+export type Address = {
+  street1: string;
+  street2: string;
+  city: string;
+  state: string;
+  postalCode: string;
+};
 
 export const PrescribeProvider = (props: PrescribeProviderProps) => {
   const { googleMapsServices } = useGoogleService();
-  const { dispatchDraftPrescriptionCreated } = usePrescribeEventDispatch();
-  const [isLoadingPrefills, setIsLoadingPrefills] = createSignal<boolean>(false);
-  const [hasCreatedPrescriptions, setHasCreatedPrescriptions] = createSignal<boolean>(false);
+  const { draftPrescriptions, setDraftPrescriptions } = useDraftPrescriptions();
+  const client = usePhotonClient();
   const [coverageOptions, setCoverageOptions] = createSignal<CoverageOption[]>([]);
   const [patientPreferredPharmacyId, setPatientPreferredPharmacyId] = createSignal<string | null>(
     null
@@ -142,10 +104,6 @@ export const PrescribeProvider = (props: PrescribeProviderProps) => {
   >();
 
   const [orderFormData, setOrderFormData] = createStore<PrescribeOrderFormData>();
-
-  const client = usePhotonClient();
-  const { draftPrescriptions, setDraftPrescriptions } = useDraftPrescriptions();
-  const [, recentOrdersActions] = useRecentOrders();
 
   const routingConstraints = createMemo((): RoutingConstraint[] => {
     return draftPrescriptions().map((prescription: Prescription) =>
@@ -191,20 +149,6 @@ export const PrescribeProvider = (props: PrescribeProviderProps) => {
     const pharmacies = await fetchPharmacies(locations[0]);
     return pharmacies;
   }
-
-  // Prefill new prescriptions based on templateIds or prescriptionIds when we get a patientId
-  createEffect(() => {
-    if (
-      // must have templateIds or prescriptionIds to create prescriptions
-      (props.templateIdsPrefill.length > 0 || props.prescriptionIdsPrefill.length > 0) &&
-      // must have a patientId
-      !!props.patientId &&
-      // must not have created prescriptions yet
-      !hasCreatedPrescriptions()
-    ) {
-      createPrescriptionsFromIds();
-    }
-  });
 
   createEffect(() => {
     if (props.patientId) {
@@ -284,44 +228,6 @@ export const PrescribeProvider = (props: PrescribeProviderProps) => {
     }
   });
 
-  async function createPrescriptionsFromIds() {
-    setHasCreatedPrescriptions(true);
-    setIsLoadingPrefills(true);
-
-    try {
-      // Create prescriptions from template ids with a few optional override values
-      if (props.templateIdsPrefill.length > 0) {
-        const dedupedTemplateIds = Array.from(new Set(props.templateIdsPrefill));
-        const templatedCreateRxList = dedupedTemplateIds.map((templateId) => ({
-          ...props.templateOverrides?.[templateId],
-          patientId: props.patientId,
-          templateId
-        }));
-
-        await createPrescriptionsOnApi(templatedCreateRxList);
-      }
-
-      // Fetch prescriptions if needed
-      if (props.prescriptionIdsPrefill.length > 0) {
-        const fetchedPrescriptions = await Promise.all(
-          props.prescriptionIdsPrefill.map(async (prescriptionId: string) => {
-            const { data } = await client.apollo.query({
-              query: GetPrescription,
-              variables: { id: prescriptionId }
-            });
-            return transformPrescriptionFormData(data?.prescription, props.patientId);
-          })
-        );
-
-        await createPrescriptionsOnApi(fetchedPrescriptions);
-      }
-    } catch (error) {
-      console.error('Error while trying to create prescriptions from prefill IDs', { error });
-    } finally {
-      setIsLoadingPrefills(false);
-    }
-  }
-
   const getPatientPreferredPharmaciesAndAddress = async (patientId: string) => {
     try {
       const response = await client.apollo.query({
@@ -360,152 +266,6 @@ export const PrescribeProvider = (props: PrescribeProviderProps) => {
     return response.data.generateCoverageOptions as CoverageOption[];
   };
 
-  const tryCreatePrescription = async (
-    prescriptionFormData: PrescriptionFormData,
-    options: TryCreatePrescriptionTemplateOptions = { showSuccessToast: true }
-  ): Promise<Prescription> => {
-    const isPrescriptionAlreadyAdded = isTreatmentInDraftPrescriptions(
-      prescriptionFormData.treatment.id,
-      draftPrescriptions()
-    );
-
-    if (isPrescriptionAlreadyAdded) {
-      triggerToast({
-        status: 'error',
-        body: 'You already have this prescription in your order. You can modify the prescription or delete it in Pending Order.'
-      });
-
-      throw new Error('Prescription already added');
-    }
-
-    const duplicateFill = recentOrdersActions.checkDuplicateFill(
-      prescriptionFormData.treatment.name
-    );
-
-    if (props.enableCombineAndDuplicate && duplicateFill) {
-      // if there's a duplicate order, check first if they want to report an issue
-      await new Promise<void>((resolve, reject) => {
-        recentOrdersActions.setIsDuplicateDialogOpen(true, duplicateFill, resolve, reject);
-      });
-    }
-
-    return await createPrescriptionOnApi(prescriptionFormData, options);
-  };
-
-  const tryUpdatePrescriptionStates = async (
-    ids: string[],
-    state: PrescriptionState
-  ): Promise<boolean> => {
-    try {
-      const res = await client.apolloClinical.mutate({
-        mutation: UpdatePrescriptionStates,
-        variables: {
-          input: {
-            ids,
-            state
-          }
-        }
-      });
-
-      return res.data.updatePrescriptionStates as boolean;
-    } catch (error) {
-      console.error('Mutation error:', error);
-      triggerToast({
-        status: 'error',
-        header: 'Error Saving Prescription(s)',
-        body: (error as Error).message
-      });
-      throw error;
-    }
-  };
-
-  const createPrescriptionOnApi = async (
-    prescriptionFormData: PrescriptionFormData,
-    options: TryCreatePrescriptionTemplateOptions = {
-      addToTemplates: false,
-      showSuccessToast: true
-    }
-  ): Promise<Prescription> => {
-    let createdPrescription: Prescription | null = null;
-
-    try {
-      const res = await client.apollo.mutate({
-        mutation: CreatePrescription,
-        variables: transformPrescriptionFormData(prescriptionFormData, props.patientId)
-      });
-      const created = res.data.createPrescription as Prescription;
-      createdPrescription = created;
-      setDraftPrescriptions((prev) => [...prev, created]);
-      dispatchDraftPrescriptionCreated(created);
-    } catch (error) {
-      console.error('Mutation error:', error);
-      triggerToast({
-        status: 'error',
-        header: 'Error Adding Prescription',
-        body: 'There was an issue adding the prescription. Please try again.'
-      });
-      throw error;
-    }
-
-    if (options?.addToTemplates && options?.catalogId != null) {
-      await createPrescriptionTemplateOnApi(
-        prescriptionFormData,
-        options.catalogId,
-        options.templateName
-      );
-
-      if (options.showSuccessToast) {
-        triggerToast({
-          status: 'success',
-          header: 'Personal Template Saved'
-        });
-      }
-    }
-
-    if (options.showSuccessToast) {
-      triggerToast({
-        status: 'success',
-        header: 'Prescription Added',
-        body: 'You can send this order or add another prescription before sending it'
-      });
-    }
-
-    return createdPrescription;
-  };
-
-  const createPrescriptionsOnApi = async (
-    prescriptions: MutationCreatePrescriptionsArgs['prescriptions']
-  ) => {
-    const res = await client.apollo.mutate({
-      mutation: CreatePrescriptions,
-      variables: { prescriptions }
-    });
-    const newRxs = res.data.createPrescriptions as Prescription[];
-    setDraftPrescriptions((prev) => [...prev, ...newRxs]);
-    newRxs.map((rx) => dispatchDraftPrescriptionCreated(rx));
-  };
-
-  const deletePrescription = (toDeleteId: string) => {
-    setDraftPrescriptions((prev) => prev.filter((rx) => rx.id !== toDeleteId));
-  };
-
-  const createPrescriptionTemplateOnApi = async (
-    prescription: PrescriptionFormData,
-    catalogId: string,
-    templateName = ''
-  ) => {
-    const res = await client.apollo.mutate({
-      mutation: CreatePrescriptionTemplate,
-      variables: {
-        ...transformPrescriptionFormData(prescription, props.patientId),
-        catalogId,
-        isPrivate: true,
-        ...(templateName ? { name: templateName } : {})
-      }
-    });
-    return res;
-  };
-
   const selectOtherCoverageOption = (value: CoverageOption) => {
     setOrderFormData('pharmacyId', value.pharmacy.id);
     setSelectedCoverageOption(value);
@@ -514,7 +274,7 @@ export const PrescribeProvider = (props: PrescribeProviderProps) => {
 
   const value = {
     // values
-    isLoadingPrefills,
+
     coverageOptions,
     routingConstraints,
     combinedRoutingConstraint,
@@ -523,9 +283,7 @@ export const PrescribeProvider = (props: PrescribeProviderProps) => {
     selectedCoverageOption,
 
     // actions
-    tryCreatePrescription, // used
-    tryUpdatePrescriptionStates,
-    deletePrescription,
+
     selectOtherCoverageOption,
     setOrderFormData
   };
@@ -539,51 +297,4 @@ export const usePrescribe = () => {
     throw new Error('usePrescribe must be used within the PrescribeProvider');
   }
   return context;
-};
-
-export const usePrescribeOptional = () => {
-  return useContext(PrescribeContext);
-};
-
-export function isTreatmentInDraftPrescriptions(
-  treatmentId: string,
-  draftedPrescriptions: { treatment: { id: string } }[]
-) {
-  return draftedPrescriptions.some((draft) => draft.treatment.id === treatmentId);
-}
-
-export type TryCreatePrescriptionTemplateOptions = {
-  addToTemplates?: boolean;
-  templateName?: string;
-  catalogId?: string;
-  showSuccessToast?: boolean;
-};
-
-export type PatientPreferredPharmacy = {
-  id: string;
-  name: string;
-};
-
-export type CoverageOption = {
-  daysSupply: number;
-  dispenseQuantity: number;
-  dispenseUnit: string;
-  id: string;
-  isAlternative: boolean;
-  paRequired: boolean;
-  prescriptionId: string;
-  price: number | null;
-  status: 'COVERED' | 'COVERED_WITH_RESTRICTIONS' | 'NOT_COVERED';
-  statusMessage: string;
-  treatment: { id: string; name: string };
-  alerts: Array<{ label: string; text: string }>;
-  pharmacy: { id: string; name: string };
-};
-
-export type Address = {
-  street1: string;
-  street2: string;
-  city: string;
-  state: string;
-  postalCode: string;
 };
